refactor : 검색 관련 관련 도메인 클래스 이동 (#86)
