+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/service/ProductPageService.java
@@ -6,16 +6,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.breaditnow.customer.domain.product.controller.req.ProductSearchRequest;
 import com.breaditnow.customer.domain.product.controller.res.HotProductPageResponse;
 import com.breaditnow.customer.domain.product.controller.res.HotProductResponse;
-import com.breaditnow.customer.domain.product.controller.res.SearchProductPageResponse;
-import com.breaditnow.customer.domain.product.controller.res.SearchProductResponse;
 import com.breaditnow.domain.domain.favorite.repository.customerproductfavorite.CustomerProductFavoriteRepository;
 import com.breaditnow.domain.domain.product.entity.Product;
 import com.breaditnow.domain.domain.product.repository.ProductRepository;
-import com.breaditnow.domain.domain.product.repository.SearchProductRepository;
-import com.breaditnow.domain.global.dto.GeoPoint;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,7 +19,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ProductPageService {
 	private final ProductRepository productRepository;
-	private final SearchProductRepository searchProductRepository;
 	private final CustomerProductFavoriteRepository favoriteRepository;
 
 	public HotProductPageResponse searchHotProducts(Long customerId, int page, int size, String sort) {
@@ -40,18 +34,4 @@ public class ProductPageService {
 		return HotProductPageResponse.of(hotProductResponses);
 	}
 
-	public SearchProductPageResponse searchProducts(Long customerId, ProductSearchRequest searchRequest) {
-		Pageable pageable = PageRequest.of(searchRequest.page(), searchRequest.size());
-		GeoPoint geoPoint = GeoPoint.of(searchRequest.latitude(), searchRequest.longitude());
-
-		Page<Product> productPage = searchProductRepository.searchProductsWithKeyword(pageable, searchRequest.sort(),
-			searchRequest.keyword(), geoPoint);
-
-		Page<SearchProductResponse> searchProductResponses = productPage.map(product -> {
-			boolean isFavorite = favoriteRepository.existsByCustomerIdAndProductId(customerId, product.getId());
-			return SearchProductResponse.of(product, isFavorite);
-		});
-
-		return SearchProductPageResponse.of(searchProductResponses);
-	}
 }
