+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/search/controller/SearchController.java
@@ -8,8 +8,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.breaditnow.common.response.ApiSuccessResponse;
+import com.breaditnow.customer.domain.product.controller.res.SearchProductPageResponse;
+import com.breaditnow.customer.domain.search.controller.req.SearchRequest;
 import com.breaditnow.customer.domain.search.controller.res.SearchAutoCompleteResponse;
-import com.breaditnow.customer.domain.search.service.AutoCompleteService;
+import com.breaditnow.customer.domain.search.controller.res.SearchBakeryPageResponse;
+import com.breaditnow.customer.domain.search.service.SearchService;
+import com.breaditnow.customer.global.security.annotation.AuthCustomer;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,7 +21,32 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/search")
 public class SearchController {
-	private final AutoCompleteService searchService;
+	private final SearchService searchService;
+
+	@GetMapping("/bakery")
+	public ApiSuccessResponse<SearchBakeryPageResponse> searchBakeries(
+		@AuthCustomer Long customerId,
+		@RequestParam(name = "page", defaultValue = "0") int page,
+		@RequestParam(name = "size", defaultValue = "10") int size,
+		@RequestParam(name = "sort", defaultValue = "latest") String sort,
+		@RequestParam(name = "keyword", defaultValue = "") String keyword,
+		@RequestParam(name = "latitude", required = false) Double latitude,
+		@RequestParam(name = "longitude", required = false) Double longitude) {
+		return ApiSuccessResponse.of(searchService.searchBakeries(customerId,
+			SearchRequest.of(page, size, sort, keyword, latitude, longitude)));
+	}
+
+	@GetMapping("/product")
+	public ApiSuccessResponse<SearchProductPageResponse> searchProducts(@AuthCustomer Long customerId,
+		@RequestParam(name = "page", defaultValue = "0") int page,
+		@RequestParam(name = "size", defaultValue = "10") int size,
+		@RequestParam(name = "sort", defaultValue = "latest") String sort,
+		@RequestParam(name = "keyword", defaultValue = "") String keyword,
+		@RequestParam(name = "latitude", required = false) Double latitude,
+		@RequestParam(name = "longitude", required = false) Double longitude) {
+		return ApiSuccessResponse.of(searchService.searchProducts(customerId,
+			SearchRequest.of(page, size, sort, keyword, latitude, longitude)));
+	}
 
 	@GetMapping("/autocomplete")
 	public ApiSuccessResponse<List<SearchAutoCompleteResponse>> searchAutoComplete(
